test(platform): isolate Clerk HTML transformer types
