+++ turbo/apps/platform/scripts/clerk-html.ts
@@ -1,17 +1,9 @@
-import { clerkJSScriptUrl } from "@clerk/shared/loadClerkJsScript";
 import type { Plugin } from "vite";
 
-import { CLERK_JS_VERSION } from "../src/lib/clerk-versions.ts";
-
-const PREVIEW_SCRIPT_URL_MARKER = "__VM0_CLERK_PREVIEW_SCRIPT_URL__";
-const PRODUCTION_SCRIPT_URL_MARKER = "__VM0_CLERK_PRODUCTION_SCRIPT_URL__";
-const SATELLITE_SCRIPT_URL_MARKER = "__VM0_CLERK_SATELLITE_SCRIPT_URL__";
-const PRODUCTION_SATELLITE_DOMAIN = "app.okou.ai";
-
-interface ClerkCoreHtmlOptions {
-  readonly previewPublishableKey: string;
-  readonly productionPublishableKey: string;
-}
+import {
+  type ClerkCoreHtmlOptions,
+  transformClerkCoreScriptUrls,
+} from "./clerk-html-transform.ts";
 
 function requiredEnvironmentValue(
   environment: Record<string, unknown>,
@@ -22,33 +14,6 @@ function requiredEnvironmentValue(
     throw new Error(`Missing ${name} environment variable`);
   }
   return value;
-}
-
-function scriptUrl(publishableKey: string, domain?: string): string {
-  return clerkJSScriptUrl({
-    __internal_clerkJSVersion: CLERK_JS_VERSION,
-    domain,
-    publishableKey,
-  });
-}
-
-export function transformClerkCoreScriptUrls(
-  html: string,
-  options: ClerkCoreHtmlOptions,
-): string {
-  return html
-    .replaceAll(
-      PREVIEW_SCRIPT_URL_MARKER,
-      scriptUrl(options.previewPublishableKey),
-    )
-    .replaceAll(
-      PRODUCTION_SCRIPT_URL_MARKER,
-      scriptUrl(options.productionPublishableKey),
-    )
-    .replaceAll(
-      SATELLITE_SCRIPT_URL_MARKER,
-      scriptUrl(options.productionPublishableKey, PRODUCTION_SATELLITE_DOMAIN),
-    );
 }
 
 export function clerkCoreHtmlPlugin(): Plugin {
